// src/app/models/v_TreatmentsHeadersFull.model.ts
export class ViewTreatmentHeaderFull {
    idTreatmentHeader: number;
    idReasonAppointment: number;
    reasonAppointment: string;
    isLaser: boolean;
    idUser: number;
    firstNameUser: string;
    idPatient: number;
    fullName: string;
    numberSessionsAssigned: number;
    numberSessionsAttended: number;
    dateStartAttention: Date;
    dateLastAttention: Date;
    attentionStatus: boolean;

}