// src/app/models/v_TreatmentsDetailsFull.model.ts
export class ViewTreatmentDetailFull {
    idTreatmentDetail: number;
    idTreatmentHeader: number;
    idDiagnostic: number;
    diagnostic: string;
    idTypeMeasure: number;
    typeMeasure: string;
    dateTreatmentDetail: Date;
    sistolica: number;
    diastolica: number;
    heartRate: number;
    evolutionRegister: string;
    pending: string;
    diagnosticComments: string;
    periodicity: number;
    idUserLogin: number;
    fullNameUser: string;

}